+++ mod-pubsub-server/src/main/java/org/folio/kafka/KafkaConfig.java
@@ -15,6 +15,8 @@ public class KafkaConfig {
   private String kafkaHost;
   @Value("${KAFKA_PORT}")
   private String kafkaPort;
+  @Value("${OKAPI_URL}")
+  private String okapiUrl;
 
   public String getKafkaHost() {
     return kafkaHost;
@@ -24,9 +26,13 @@ public class KafkaConfig {
     return kafkaPort;
   }
 
+  public String getOkapiUrl() {
+    return okapiUrl;
+  }
+
   public Map<String, String> getProducerProps() {
     Map<String, String> producerProps = new HashMap<>();
-    producerProps.put(ProducerConfig.BOOTSTRAP_SERVERS_CONFIG, getUrl());
+    producerProps.put(ProducerConfig.BOOTSTRAP_SERVERS_CONFIG, getKafkaUrl());
     producerProps.put(ProducerConfig.ENABLE_IDEMPOTENCE_CONFIG, "true");
     producerProps.put(ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG, "org.apache.kafka.common.serialization.StringSerializer");
     producerProps.put(ProducerConfig.VALUE_SERIALIZER_CLASS_CONFIG, "org.apache.kafka.common.serialization.StringSerializer");
@@ -35,7 +41,7 @@ public class KafkaConfig {
 
   public Map<String, String> getConsumerProps() {
     Map<String, String> consumerProps = new HashMap<>();
-    consumerProps.put(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG, getUrl());
+    consumerProps.put(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG, getKafkaUrl());
     consumerProps.put(ConsumerConfig.ENABLE_AUTO_COMMIT_CONFIG, "true");
     consumerProps.put(ConsumerConfig.AUTO_COMMIT_INTERVAL_MS_CONFIG, "1000");
     consumerProps.put(ConsumerConfig.KEY_DESERIALIZER_CLASS_CONFIG, "org.apache.kafka.common.serialization.StringDeserializer");
@@ -43,7 +49,7 @@ public class KafkaConfig {
     return consumerProps;
   }
 
-  public String getUrl() {
+  public String getKafkaUrl() {
     return kafkaHost + ":" + kafkaPort;
   }
 }