+++ mod-pubsub-server/src/main/java/org/folio/services/StartupService.java
@@ -1,0 +1,12 @@
+package org.folio.services;
+
+/**
+ * Startup Service Interface
+ */
+public interface StartupService {
+
+  /**
+   * Initializes all registered active subscribers
+   */
+  void initSubscribers();
+}